astro-project/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
duckdb==0.10.3
tabulate==0.9.0
apache-airflow-providers-google